dont define instances alongside class definitions
