+++ base/shared/src/main/scala/scalaz/core/eq.scala
@@ -7,10 +7,6 @@ trait EqClass[A] {
   def equal(first: A, second: A): Boolean
 }
 
-trait EqInstances {
-  implicit final val voidEq: Eq[Void] = instanceOf[EqClass[Void]]((a, b) => a.absurd)
-}
-
 trait EqSyntax {
   implicit final class ToEqOps[A](a: A) {
     private[core] type equal
