+++ base/shared/src/main/scala/scalaz/ct/bifunctor.scala
@@ -1,8 +1,6 @@
 package scalaz
 package ct
 
-import scala.Tuple2
-
 import scala.language.experimental.macros
 
 /** A typeclass for types which are (covariant) [[Functor]]s in both type parameters.
@@ -47,15 +45,6 @@ object BifunctorClass {
 
     def map[B, BB](ma: F[A, B])(f: B => BB): F[A, BB] = F.rmap(ma)(f)
   }
-}
-
-trait BifunctorInstances {
-
-  implicit final val tuple2Bifunctor: Bifunctor[Tuple2] =
-    instanceOf(new BifunctorClass[Tuple2] with BifunctorClass.DeriveBimap[Tuple2] {
-      def lmap[A, B, S](fab: (A, B))(f: A => S): (S, B) = fab.copy(_1 = f(fab._1))
-      def rmap[A, B, T](fab: (A, B))(f: B => T): (A, T) = fab.copy(_2 = f(fab._2))
-    })
 }
 
 trait BifunctorSyntax {
