+++ base/shared/src/main/scala/scalaz/ct/bind.scala
@@ -30,8 +30,6 @@ trait BindFunctions {
   def flatMap[M[_], A, B](ma: M[A])(f: A => M[B])(implicit M: Bind[M]): M[B] = M.flatMap(ma)(f)
 }
 
-trait BindInstances {}
-
 trait BindSyntax {
   implicit final class ToBindOps[M[_], A](ma: M[A]) {
     def flatMap[B](f: A => M[B])(implicit ev: Bind[M]): M[B] = macro meta.Ops.i_1
