+++ base/shared/src/main/scala/scalaz/ct/cobind.scala
@@ -1,8 +1,6 @@
 package scalaz
 package ct
 
-import scala.{ List, Option, Some }
-
 import scala.language.experimental.macros
 
 trait CobindClass[F[_]] extends FunctorClass[F] {
@@ -24,22 +22,6 @@ object CobindClass {
 
   trait Alt[D <: Alt[D]]
 
-}
-
-trait CobindInstances {
-  implicit val optionCobind: Cobind[Option] = instanceOf(new CobindClass.DeriveCojoin[Option] {
-    override def map[A, B](fa: Option[A])(f: A => B): Option[B] = fa.map(f)
-
-    override def cobind[A, B](fa: Option[A])(f: Option[A] => B): Option[B] =
-      Some(f(fa))
-  })
-
-  implicit val listCobind: Cobind[List] = instanceOf(new CobindClass.DeriveCojoin[List] {
-    override def map[A, B](fa: List[A])(f: A => B): List[B] = fa.map(f)
-
-    override def cobind[A, B](fa: List[A])(f: List[A] => B): List[B] =
-      List(f(fa))
-  })
 }
 
 trait CobindSyntax {
